+++ src/constants/data.ts
@@ -5,7 +5,7 @@ interface MenuI {
   link: string;
   highlight?: boolean;
 }
-const MenuList: readonly MenuI[] = [
+export const menuList: readonly MenuI[] = [
   { text: '공지사항', link: URL.menu.notice },
   { text: '자주 묻는 질문', link: URL.menu.faq },
   { text: '포즈피커 공식 SNS', link: URL.instagram, highlight: true },
@@ -15,4 +15,20 @@ const MenuList: readonly MenuI[] = [
   { text: '개인정보 처리방침', link: URL.menu.privacy },
 ] as const;
 
-export default MenuList;
+export const peopleCountList = [
+  { text: '전체', value: 0 },
+  { text: '1인', value: 1 },
+  { text: '2인', value: 2 },
+  { text: '3인', value: 3 },
+  { text: '4인', value: 4 },
+  { text: '5인+', value: 5 },
+];
+
+export const frameCountList = [
+  { text: '전체', value: 0 },
+  { text: '1컷', value: 1 },
+  { text: '3컷', value: 3 },
+  { text: '4컷', value: 4 },
+  { text: '6컷', value: 6 },
+  { text: '8컷+', value: 8 },
+];
