--- src/constants/filterList.ts
@@ -1,17 +1,0 @@
-export const peopleCountList = [
-  { text: '전체', value: 0 },
-  { text: '1인', value: 1 },
-  { text: '2인', value: 2 },
-  { text: '3인', value: 3 },
-  { text: '4인', value: 4 },
-  { text: '5인+', value: 5 },
-];
-
-export const frameCountList = [
-  { text: '전체', value: 0 },
-  { text: '1컷', value: 1 },
-  { text: '3컷', value: 3 },
-  { text: '4컷', value: 4 },
-  { text: '6컷', value: 6 },
-  { text: '8컷+', value: 8 },
-];
